+++ client/src/utility/redirectUrl.ts
@@ -1,7 +1,11 @@
+import { urlParser } from '.';
+
 export const redirectUrl = (url: string, sameTab: boolean) => {
+  const parsedUrl = urlParser(url)[1];
+
   if (sameTab) {
-    document.location.replace(url);
+    document.location.replace(parsedUrl);
   } else {
-    window.open(url);
+    window.open(parsedUrl);
   }
 };